+++ net/can/j1939/j1939-priv.h
@@ -174,7 +174,7 @@ static inline void j1939_ecu_remove_sa(struct j1939_ecu *ecu)
 	write_unlock_bh(&ecu->priv->lock);
 }
 
-int j1939_name_to_sa(name_t name, int ifindex);
+u8 j1939_name_to_sa(name_t name, int ifindex);
 struct j1939_ecu *j1939_ecu_find_by_addr(u8 sa, int ifindex);
 struct j1939_ecu *j1939_ecu_find_by_name(name_t name, int ifindex);
 /* find_by_name, with kref & read_lock taken */
